# swagger.yaml
openapi: 3.0.2
info:
    title: Property management
    version: 1.0.0
    description: Property management API docs
    tags:
        -
            name: properties
            description: Operations related to properties
        -
            name: sessions
            description: Operations related to sessions
paths:
    /properties:
        get:
            tags:
                - properties
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/getPropertiesResponse'
                            examples:
                                getPropertiesExample:
                                    value:
                                        -
                                            createdAt: '2022-11-01 00:00:00'
                                            name: string
                                            location: string
                    description: Returns an array of properties
                '403':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Required:
                                    value: A token is required for authentication
                    description: Token required
            security:
                -
                    ApiKeyAuth: []
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            type: string
                        examples:
                            Valid data:
                                value:
                                    name: Property 222
                                    location: Somewhere in Estonia aaa
                required: true
            tags:
                - properties
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Created:
                                    value:
                                        createdAt: '2022-11-17 08:39:37'
                                        name: Property 2
                                        _id: 6375d729f0f5667c4857d4bd
                                        location: Somewhere in Estoniaaaa
                                        __v: 0
                    description: returns Created
                '400':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Required:
                                    value: All input is required
                    description: Invalid or missing data
                '409':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Duplicate:
                                    value: Property Already Exist!
                    description: Duplicate
            security:
                -
                    ApiKeyAuth: []
        delete:
            tags:
                - properties
            responses:
                '204':
                    description: Success
                '404':
                    description: no Property with given id found
            security:
                -
                    ApiKeyAuth: []
    /sessions:
        post:
            requestBody:
                content:
                    application/json:
                        schema:
                            type: string
                        examples:
                            Sessions_example:
                                value:
                                    email: urmas.muldmaa@vikk.ee
                                    password: '1234567'
                required: true
            tags:
                - sessions
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Success:
                                    value:
                                        first_name: Urmas
                                        last_name: Muldmaa
                                        _id: 637551d00d5e26274ca15b8a
                                        email: urmas.muldmaa@vikk.ee
                                        password: $2a$10$NOfG37XC/ze4QMlgTakLCuFgbNpck/YUOJe35u.tIUdEo8wyPV/Bq
                                        __v: 0
                                        token: >-
                                            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjM3NTUxZDAwZDVlMjYyNzRjYTE1YjhhIiwiZW1haWwiOiJ1cm1hcy5tdWxkbWFhQHZpa2suZWUiLCJpYXQiOjE2Njg2NjYwMjEsImV4cCI6MTY2ODY3MzIyMX0.Ygqo7xqcS8PTzcKse0MSQZksRSzBuVMRsLPbEjHZdWc
                    description: User login successful
                '400':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Invalid:
                                    value: Invalid Credentials
                    description: Invalid credentials
                '411':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Required:
                                    value: All input is required
                    description: Input required
components:
    schemas:
        getPropertiesResponse:
            description: ''
            type: array
            items:
                $ref: '#/components/schemas/getPropertiesResponse'
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            description: API Key
            name: x-access-token
            in: header